[UPDATE] 스프링 시큐리티 적용 완료 #[MORE] 2차 구현 ## BE - 회원가입시 아이디, 이름 체크하기 (BE) - 로그인 시 핸들러 생성(성공, 실패) - 로그아웃시 핸들러 생성(성공, 실패) - responseEntity 응답 구조 수정
## FE
- axios post 할때 아이디, 이름 체크하기 (FE)
- axios 로 회원 가입할때 alert 창 수정
- 로그인 시 alert 창 생성
- 로그아웃 시 alert 창 생성

+++ src/main/java/com/example/myspringbootproject/domain/user/controller/UserController.java
@@ -20,10 +20,10 @@ public class UserController {
     @PostMapping
     public ResponseEntity<?> signUpUser(@RequestBody UserDTO userDTO){
         try{
-            List<Integer> userId = new ArrayList<>();
+            List<Long> userId = new ArrayList<>();
             userId.add(memberService.signUp(userDTO));
 
-            ResponseDTO<Integer> response = ResponseDTO.<Integer>builder().data(userId).build();
+            ResponseDTO<Long> response = ResponseDTO.<Long>builder().data(userId).build();
 
             // TODO ResponseEntity.ok(url).body(response) 로 변환해 보기
             return ResponseEntity.status(HttpStatus.CREATED).body(response);
